+++ mycarehub/content/tests/test_serializers.py
@@ -1,7 +1,7 @@
 import pytest
 from django.utils import timezone
 from model_bakery import baker
-from wagtail.core.models import Page
+from wagtail.models import Page
 
 from mycarehub.content.models import (
     Author,
