+++ mycarehub/home/models.py
@@ -1,6 +1,9 @@
-from wagtail.admin.edit_handlers import FieldPanel
-from wagtail.core.fields import RichTextField
-from wagtail.core.models import Page
+from django.db import models
+from wagtail.admin.panels import FieldPanel
+from wagtail.fields import RichTextField
+from wagtail.models import Page
+
+from mycarehub.common.models import Organisation
 
 
 class HomePage(Page):
@@ -9,6 +12,13 @@ class HomePage(Page):
     """
 
     body = RichTextField(blank=True)
+    organisation = models.ForeignKey(
+        Organisation,
+        on_delete=models.PROTECT,
+        null=True,
+        blank=True,
+        related_name="%(app_label)s_%(class)s_related",
+    )
 
     content_panels = Page.content_panels + [
         FieldPanel("body", classname="full"),
