+++ mycarehub/home/tests/test_models.py
@@ -1,5 +1,5 @@
 import pytest
-from wagtail.core.models import Page
+from wagtail.models import Page
 
 from mycarehub.home.models import HomePage
