+++ mycarehub/templates/content/content_item.html
@@ -2,7 +2,7 @@
 
 {% load wagtailcore_tags %}
 {% load wagtailcore_tags wagtailimages_tags %}
-{% load wagtailvideos_tags %}
+
 
 {% block page %}{% endblock page %}
