+++ requirements/base.txt
@@ -10,8 +10,8 @@ django-crispy-forms~=1.13.0  # https://github.com/django-crispy-forms/django-cri
 django-environ~=0.8.1  # https://github.com/joke2k/django-environ
 django-filter>=2.2,<3.0 # this constraint is required for wagtail
 django-mjml[requests]~=0.11.0
-django-taggit~=1.5.1
-django-modelcluster~=5.2
+django-taggit~=2.0
+django-modelcluster~=6.0
 django-model-utils~=4.2.0  # https://github.com/jazzband/django-model-utils
 django-oauth-toolkit~=1.5.0
 django-phonenumber-field~=5.2.0
@@ -39,14 +39,13 @@ rcssmin~=1.0.6  # https://github.com/ndparker/rcssmin
 requests~=2.26.0
 uvicorn[standard]~=0.15.0  # https://github.com/encode/uvicorn
 whitenoise~=5.3.0  # https://github.com/evansd/whitenoise
-wagtail~=2.15.1
+wagtail~=4.1.1
 Wand~=0.6.7
 opencv_python~=4.5.4.58
 rustface~=0.1.0
 wagtailvideos~=2.10.9
-wagtailmedia~=0.8.0
+wagtailmedia~=0.12.0
 wagtail-readinglevel~=3.5.0
 wagtailfontawesome~=1.2.1
-wagtail-quick-create~=1.0.7
-wagtail-automatic-redirects~=1.1.5
+wagtail-quick-create~=2.0.0
 django-nested-admin==3.4.0
